requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
# google-cloud-aiplatform>=1.38
#fastapi
#uvicorn
sentence-transformers
langchain_groq